Add request timeout to handle crunchy pods hanging
Refs: https://github.ibm.com/velox/platform/issues/8286

- I was unable to replicate the client scenario but added a potential fix given the known information.
Instead of adding --request-timeout command I added --pod-running-timeout which waits the chosen length of time until at least one pod is running, in this case 1 minute. While comparing the files generated from before and after adding the new command I did notice that some of the files had been extended but none of the information was missing there was just more activity to report I think due to the extension of running time from the command. I found this particularly within the events log and some of the pod logs.